+++ index.html
@@ -8,9 +8,20 @@
       <script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.14.3/umd/popper.min.js" integrity="sha384-ZMP7rVo3mIykV+2+9J3UJ46jBk0WLaUAdn689aCwoqbBJiSnjAK/l8WvCWPIPm49" crossorigin="anonymous"></script>
       <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.1.3/js/bootstrap.min.js" integrity="sha384-ChfqqxuZUCnJSK3+MXmPNIyE6ZbWh2IMqE241rYiqJxyMiZ6OW/JmZQ5stwEULTy" crossorigin="anonymous"></script>
       <link rel="stylesheet" href="css/main.css">
-      <script defer src="js/main.js"></script>
+      <script defer src="js/main.js" type="module"></script>
       <title>Tic Tac Toe</title>
   </head>
   <body class="pl-xl-5 pr-xl-5 bg-lgrey container">
+    <div id="board" class="row">
+      <div class="col-4"><button class="btn btn-primary" data-number="1">-</button></div>
+      <div class="col-4"><button class="btn btn-primary" data-number="2">-</button></div>
+      <div class="col-4"><button class="btn btn-primary" data-number="3">-</button></div>
+      <div class="col-4"><button class="btn btn-primary" data-number="4">-</button></div>
+      <div class="col-4"><button class="btn btn-primary" data-number="5">-</button></div>
+      <div class="col-4"><button class="btn btn-primary" data-number="6">-</button></div>
+      <div class="col-4"><button class="btn btn-primary" data-number="7">-</button></div>
+      <div class="col-4"><button class="btn btn-primary" data-number="8">-</button></div>
+      <div class="col-4"><button class="btn btn-primary" data-number="9">-</button></div>
+    </div>
   </body>
 </html>
